cats-api/api/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
astroid==3.3.3
black==24.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.8
dnspython==2.6.1
docker==7.1.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
h11==0.14.0
idna==3.10
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql==0.0.3
mysqlclient==2.2.4
packaging==24.1
pathspec==0.12.1
peewee==3.17.6
platformdirs==4.3.6
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pylint==3.3.0
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
starlette==0.38.6
tomli==2.0.1
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
